backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hboyjd-srm-backend"
version = "0.1.0"
description = "欧阳聚德 SRM 后端骨架"
requires-python = ">=3.12"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "sqlmodel",
  "alembic",
  "psycopg[binary]",
  "redis",
  "python-jose[cryptography]",
  "passlib[bcrypt]",
  "pydantic-settings",
  "python-multipart",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "httpx",
  "pytest-asyncio",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.setuptools.packages.find]
include = ["app*"]
